fix : 권한 추가

+++ src/main/java/com/todo/mirujima_be/note/service/NoteService.java
@@ -1,5 +1,8 @@
 package com.todo.mirujima_be.note.service;
 
+import static com.todo.mirujima_be.auth.util.AuthUtil.checkAuthority;
+
+import com.todo.mirujima_be.common.contant.MirujimaConstants;
 import com.todo.mirujima_be.common.exception.AlertException;
 import com.todo.mirujima_be.note.dto.NotePageCollection;
 import com.todo.mirujima_be.note.dto.request.NoteListRequest;
@@ -19,46 +22,58 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
 public class NoteService {
 
-    private final NoteRepository noteRepository;
-    private final TodoRepository todoRepository;
+  private final NoteRepository noteRepository;
+  private final TodoRepository todoRepository;
 
-    public NotePageCollection getNoteList(NoteListRequest noteListRequest) {
-        var noteCount = noteRepository.getNoteCount(noteListRequest);
-        var noteIds = noteRepository.getNoteIdList(noteListRequest);
-        var notes = noteRepository.findAllById(noteIds);
-        var lastSeenTodoId = notes.stream().mapToLong(Note::getId).min().orElse(0L);
-        return NotePageCollection.builder()
-                .lastSeenId(lastSeenTodoId)
-                .totalCount(noteCount.intValue())
-                .notes(notes.stream().map(NoteResponse::of).toList())
-                .build();
-    }
+  public NotePageCollection getNoteList(NoteListRequest noteListRequest) {
+    var noteCount = noteRepository.getNoteCount(noteListRequest);
+    var noteIds = noteRepository.getNoteIdList(noteListRequest);
+    var notes = noteRepository.findAllById(noteIds);
+    var lastSeenTodoId = notes.stream().mapToLong(Note::getId).min().orElse(0L);
+    return NotePageCollection.builder()
+        .lastSeenId(lastSeenTodoId)
+        .remainingCount(noteCount.intValue() - notes.size())
+        .notes(notes.stream().map(NoteResponse::of).toList())
+        .build();
+  }
 
-    public NoteResponse getNoteDetail(long noteId) {
-        var note = noteRepository.findById(noteId).orElseThrow(() -> new AlertException("노트가 존재하지 않습니다"));
-        return NoteResponse.of(note);
-    }
+  public NoteResponse getNoteDetail(long noteId) {
+    var note = noteRepository.findById(noteId).orElseThrow(() -> new AlertException("노트가 존재하지 않습니다"));
+    checkAuthority(MirujimaConstants.Note.Note, note);
+    return NoteResponse.of(note);
+  }
 
-    @Transactional
-    public NoteResponse registerNote(NoteRegRequest noteRegRequest) {
-        var note = Note.from(noteRegRequest);
-        var todo = todoRepository.findById(noteRegRequest.getTodoId())
-                .orElseThrow(() -> new AlertException("할일이 존재하지 않습니다."));
-        note.setTodo(todo);
-        noteRepository.save(note);
-        return NoteResponse.of(note);
+  @Transactional
+  public NoteResponse registerNote(NoteRegRequest noteRegRequest) {
+    var note = Note.from(noteRegRequest);
+    var todo = todoRepository.findById(noteRegRequest.getTodoId())
+        .orElseThrow(() -> new AlertException("할일이 존재하지 않습니다."));
+    if (todo.getNote() != null) {
+      throw new AlertException("이미 노트가 존재합니다.");
     }
+    checkAuthority(MirujimaConstants.Todo.Todo, todo);
+    note.setTodo(todo);
+    noteRepository.save(note);
+    return NoteResponse.of(note);
+  }
 
-    @Transactional
-    public NoteResponse modifyNote(long id, NoteModRequest noteModRequest) {
-        var note = noteRepository.findById(id).orElseThrow(() -> new AlertException("노트가 존재하지 않습니다"));
-        note.modifyTo(noteModRequest);
-        return NoteResponse.of(note);
-    }
+  @Transactional
+  public NoteResponse modifyNote(long id, NoteModRequest noteModRequest) {
+    var note = noteRepository.findById(id).orElseThrow(() -> new AlertException("노트가 존재하지 않습니다"));
+    note.modifyTo(noteModRequest);
+    return NoteResponse.of(note);
+  }
 
-    @Transactional
-    public void deleteNote(long noteId) {
-        noteRepository.deleteById(noteId);
+  @Transactional
+  public void deleteNote(long noteId) {
+    var note = noteRepository.findById(noteId).orElseThrow(() -> new AlertException("노트가 존재하지 않습니다"));
+    checkAuthority(MirujimaConstants.Note.Note, note);
+    var todo = note.getTodo();
+    note.setTodo(null);
+    if (todo != null) {
+      todo.setNote(null);
     }
+    noteRepository.deleteById(noteId);
+  }
 
 }
